test: skew testing against 1.0, consolidate testing
Skew testing for the operator used special code because the bundle
format was new and not supported by previous releases at the time of
writing the test. Now we can test against 1.0 and use the normal skew
testing suite also for OLM.

Deployments through the operator were tested with a subset of the full
test suites. As operator and OLM are now more important, it seems
prudent to test those deployments more throroughly. If that turns out
to increase the overall test time too much, then we might remove some
of the other deployment cases, like lvm-production and
direct-production.

Multiple different framework.Framework instances were create before
some of the tests, which created unnecessary work and namespaces. Now
they share the same instance.